// src/interfaces/enums/permissions.enum.ts
export enum EPermissions {
  // Users
  ADMIN_ROOT = 'admin.access.root',
  ADMIN_ACCESS_USERS_LIST = 'admin.access.users.list',
  ADMIN_ACCESS_USERS_READ = 'admin.access.users.read',
  ADMIN_ACCESS_USERS_CREATE = 'admin.access.users.create',
  ADMIN_ACCESS_USERS_UPDATE = 'admin.access.users.update',
  ADMIN_ACCESS_USERS_UPDATE_ID = 'admin.access.user.update.id',
  // Roles
  ADMIN_ACCESS_ROLES_LIST = 'admin.access.roles.list',
  ADMIN_ACCESS_ROLES_READ = 'admin.access.roles.read',
  ADMIN_ACCESS_ROLES_CREATE = 'admin.access.roles.create',
  ADMIN_ACCESS_ROLES_UPDATE = 'admin.access.roles.update',
  // Permissions
  ADMIN_ACCESS_PERMISSIONS_LIST = 'admin.access.permissions.list',
  ADMIN_ACCESS_PERMISSIONS_READ = 'admin.access.permissions.read',
  ADMIN_ACCESS_PERMISSIONS_CREATE = 'admin.access.permissions.create',
  ADMIN_ACCESS_PERMISSIONS_UPDATE = 'admin.access.permissions.update',
  // Products
  ADMIN_ACCESS_PRODUCTS_LIST = 'admin.access.products.list',
  ADMIN_ACCESS_PRODUCTS_READ = 'admin.access.products.read',
  ADMIN_ACCESS_PRODUCTS_CREATE = 'admin.access.products.create',
  ADMIN_ACCESS_PRODUCTS_UPDATE = 'admin.access.products.update'
}