// src/interfaces/enums/error-message.enum.ts
export enum ErrorMessage {
  // Common
  INTERNAL_SERVER = 'Something went wrong!',
  REQUEST_TIMEOUT = 'Request timeout',
  NOT_FOUND = 'Not found!',
  ACCESS_TOKEN_EXPIRED = 'Access token has expired!',
  FOREIGN_KEY_CONFLICT = 'Foreign key conflict!',
  INVALID_CREDENTIALS = 'Invalid credentials!',
  INVALID_TOKEN = 'Invalid token!',
  REFRESH_TOKEN_EXPIRED = 'Refresh token has expired!',
  PERMISSION_EXISTS = `There's a permission with slug`,
  ID_NOT_EMPTY = 'ID cannot be empty!',
  FULL_NAME_MAX_LENGTH = 'Full name is too long!',
  PHONE_EXISTS = 'Phone exists!',
  EMAIL_EXISTS = 'Email exists!',
  ROLE_EXISTS = 'Role exists!',
  VALID_NUMBER = 'VALID_NUMBER',
  ID_NOT_FOUND = 'ID_NOT_FOUND',
  INVALID_ACTIVE = 'INVALID_ACTIVE',
  ID_ARRAY = 'ID_ARRAY',
  INVALID_ID = 'INVALID_ID',
  // Auth
  UNAUTHORIZED = 'UNAUTHORIZED',
  SOURCE_TOKEN_EXPIRED = 'SOURCE_TOKEN_EXPIRED',
  VERIFY_TOKEN_ERROR = 'VERIFY_TOKEN_ERROR',
  ACTIVATED_EMAIL = 'ACTIVATED_EMAIL',
  BLOCKED_EMAIL = 'BLOCKED_EMAIL',
  BLOCKED_USER = 'BLOCKED_USER',
  // User
  USER_EXISTS = 'USER_EXISTS',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  EMAIL_NOT_FOUND = 'EMAIL_NOT_FOUND',
  EMAIL_NOT_EMPTY = 'EMAIL_NOT_EMPTY',
  VERIFY_CODE_RESET_PASS_NOT_EMPTY = 'VERIFY_CODE_RESET_PASS_NOT_EMPTY',
  CODE_EXPIRED = 'CODE_EXPIRES',
  INVALID_CODE = 'INVALID_CODE',
  VERIFY_CODE_RESET = 'VERIFY_CODE_RESET',
  // Role
  ROLE_DOES_NOT_EXISTS = 'ROLE_DOES_NOT_EXISTS',
  INACTIVE_USER = 'INACTIVE_USER',
  EMAIL_PHONE_NOT_EMPTY = 'EMAIL_PHONE_NOT_EMPTY',
  PermissionExists = 'PERMISSION_EXISTS',
  RoleExists = 'ROLE_EXISTS',
  InvalidCurrentPassword = 'INVALID_CURRENT_PASSWORD',
  InvalidCredentials = 'INVALID_CREDENTIALS',
  ForeignKeyConflict = 'FOREIGN_KEY_CONFLICT',
  // Assets
  PHOTO_NOT_FOUND = 'PHOTO_NOT_FOUND',
  SCREEN_SHOTS_NOT_FOUND = 'SCREEN_SHOTS_NOT_FOUND',
  // Products
  PRODUCT_NOT_FOUND = 'PRODUCT_NOT_FOUND',
  // Category
  CATEGORY_NOT_FOUND = 'CATEGORY_NOT_FOUND',
  // Coupon
  COUPON_FAIL = 'COUPON_FAIL',
  // Order
  ORDER_NOT_FOUND = 'ORDER_NOT_FOUND'
}