// src/interfaces/message.success.ts
export enum MessageSuccesses {
  VERIFY_SUCCESS = 'Verify account successfully!',
  VERIFY_CODE_SUCCESS = 'Verify code successfully!',
  REFRESH_VERIFY_SUCCESS = 'Send mail successfully! Please check mail and verify account',
  SEND_MAIL_CODE_SUCCESS = `We're sending a verification code to your email.`,
  DELETE_SUCCESS = 'Delete successfully!',
  DELETE_SUBJECT_SUCCESS = 'Delete subject successfully!',
  SUCCESS = 'Successfully!'
}